+++ tests/one_dim/test_system.py
@@ -1,11 +1,15 @@
 """Validate the functionality of the `gspits.one_dim.system` module."""
 
 import numpy as np
+import pytest
 from attr import dataclass
+from hypothesis import given
+from hypothesis import strategies as stg
 from numba import njit
 from numpy import pi
 
-from gspits.one_dim import ExternalPotential, Hamiltonian
+from gspits import Mesh
+from gspits.one_dim import BlochState, ExternalPotential, Hamiltonian
 
 
 @njit
@@ -36,3 +40,31 @@ def test_hamiltonian():
     hamiltonian = _TestHamiltonian()
     assert isinstance(hamiltonian.external_potential, ExternalPotential)
     assert isinstance(hamiltonian, Hamiltonian)
+
+
+@given(
+    lower_bound=stg.floats(min_value=-2, max_value=-1, allow_nan=False),
+    upper_bound=stg.floats(min_value=1, max_value=2, allow_nan=False),
+    num_segments=stg.integers(min_value=128, max_value=512),
+    wave_vector=stg.floats(
+        min_value=-3 * pi, max_value=3 * pi, allow_nan=False
+    ),
+)
+def test_bloch_state(
+    lower_bound: float,
+    upper_bound: float,
+    num_segments: int,
+    wave_vector: float,
+):
+    """Check the basic functionality of `BlochState` instances."""
+    mesh = Mesh(
+        lower_bound=lower_bound,
+        upper_bound=upper_bound,
+        num_segments=num_segments,
+    )
+    bloch_state = BlochState.plane_wave(mesh, wave_vector)
+    # A plane wave is normalized, and its norm must be approximately one.
+    assert bloch_state.norm == pytest.approx(1, abs=mesh.step_size)
+    assert bloch_state.norm == pytest.approx(
+        bloch_state.periodic_component.norm, abs=1e-8
+    )